gpu: move device-function resolution and splicing into its own unit
Finding the implementation behind a `module procedure`, loading the
submodule that carries it, marking which regions device code runs in,
deciding whether a callee has to be spliced into the kernel body, and
doing the splice, are one job: getting a device callee into a shape the
kernel can hold. Nothing else in the pass asks about it.

Pure code motion: the definitions are unchanged, only their home is.
